# saas-app-ui/src/app/api-services/models/page-response-user-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { UserResponse } from '../models/user-response';
export interface PageResponseUserResponse {
  content?: Array<UserResponse>;
  first?: boolean;
  hasNext?: boolean;
  hasPrevious?: boolean;
  last?: boolean;
  page?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
}