# saas-app/src/main/java/com/amoutech/saas/responses/UserResponse.java
package com.amoutech.saas.responses;

import com.amoutech.saas.entities.UserRole;
import lombok.*;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class UserResponse {

    private String id;
    private String username;
    private String email;
    private String password;
    private String firstName;
    private String lastName;
    private UserRole role;
}